Create Enhancing Environmental Monitoring through Advanced Object  Detection in Satellite Imagery
In this project, we aim to develop an advanced object detection model using a dataset of snooker balls positioned on a snooker board (https://universe.roboflow.com/liangdianzhong/-qvdww/dataset/3). This setup provides a simplified yet effective environment to train our model to recognize and localize distinct objects within a confined setting, simulating the complexities of detecting small, scattered objects in larger images, such as animals or vehicles in environment.