Create Predictive Analytics and Machine Learning Project Report
For this project I would be using the concepts of predictive analysis and machine learning 
to carry out a customer segmentation of a dataset containing data associated to an online 
retailer’s 2 years of sales information. This dataset was accessed through ‘UCI Machine 
Learning Repository’ and consists of an Excell file having two sheets corresponding to 
different years: ‘Year 2009-2010’ and ‘Year 2010-2011.’ (UCI Machine Learning 
Repository. (2019). ‘Copy of online_retail_II.xlsx. [Dataset].) 
The goal of such a segmentation is to identify those segments which are more likely to be 
profitable or that which have a higher growth potential for the retail store.